+++ node_helper.js
@@ -7,6 +7,7 @@
 
 const Log = require('logger')
 const NodeHelper = require('node_helper')
+const undici = require('undici')
 
 module.exports = NodeHelper.create({
   start: function () {
@@ -15,9 +16,17 @@ module.exports = NodeHelper.create({
 
   async getData(payload) {
     try {
+      if (payload.allowSelfSignedCerts == true && payload.url.startsWith('https')) {
+        var agent = new undici.Agent({ connect: {rejectUnauthorized: false,},})
+        Log.debug(`${payload.url} - insecure`)
+      } else {
+        agent = new undici.Agent({ connect: {rejectUnauthorized: true,},})
+        Log.debug(`${payload.url} - secure`)
+      }
       const url = payload.url
       const response = await fetch(url, {
         method: 'GET',
+        dispatcher: agent,
       })
       var data = await response.text()
       data = data.replace(/\n+$/, '')